+++ sshd-common/src/main/java/org/apache/sshd/common/config/keys/loader/pem/ECDSAPEMResourceKeyPairParser.java
@@ -51,6 +51,7 @@ import org.apache.sshd.common.util.security.SecurityUtils;
 
 /**
  * @author <a href="mailto:dev@mina.apache.org">Apache MINA SSHD Project</a>
+ * @see    <a href="https://tools.ietf.org/html/rfc5915">RFC 5915</a>
  */
 public class ECDSAPEMResourceKeyPairParser extends AbstractPEMResourceKeyPairParser {
     public static final String BEGIN_MARKER = "BEGIN EC PRIVATE KEY";
@@ -77,7 +78,32 @@ public class ECDSAPEMResourceKeyPairParser extends AbstractPEMResourceKeyPairPar
             FilePasswordProvider passwordProvider,
             InputStream stream, Map<String, String> headers)
             throws IOException, GeneralSecurityException {
-        Map.Entry<ECPublicKeySpec, ECPrivateKeySpec> spec = decodeECPrivateKeySpec(stream, false);
+
+        KeyPair kp = parseECKeyPair(stream, false);
+        return Collections.singletonList(kp);
+    }
+
+    public static KeyPair parseECKeyPair(
+            InputStream inputStream, boolean okToClose)
+            throws IOException, GeneralSecurityException {
+        try (DERParser parser = new DERParser(NoCloseInputStream.resolveInputStream(inputStream, okToClose))) {
+            return parseECKeyPair(null, parser);
+        }
+    }
+
+    /**
+     * @param  curve                    The {@link ECCurves curve} represented by this data (in case it was optional and
+     *                                  somehow known externally) if {@code null} then it is assumed to be part of the
+     *                                  parsed data. then it is assumed to be part of the data.
+     * @param  parser                   The {@link DERParser} for the data
+     * @return                          The parsed {@link KeyPair}
+     * @throws IOException              If failed to parse the data
+     * @throws GeneralSecurityException If failed to generate the keys
+     */
+    public static KeyPair parseECKeyPair(ECCurves curve, DERParser parser)
+            throws IOException, GeneralSecurityException {
+        ASN1Object sequence = parser.readObject();
+        Map.Entry<ECPublicKeySpec, ECPrivateKeySpec> spec = decodeECPrivateKeySpec(curve, sequence);
         if (!SecurityUtils.isECCSupported()) {
             throw new NoSuchProviderException("ECC not supported");
         }
@@ -85,16 +111,15 @@ public class ECDSAPEMResourceKeyPairParser extends AbstractPEMResourceKeyPairPar
         KeyFactory kf = SecurityUtils.getKeyFactory(KeyUtils.EC_ALGORITHM);
         ECPublicKey pubKey = (ECPublicKey) kf.generatePublic(spec.getKey());
         ECPrivateKey prvKey = (ECPrivateKey) kf.generatePrivate(spec.getValue());
-        KeyPair kp = new KeyPair(pubKey, prvKey);
-        return Collections.singletonList(kp);
+        return new KeyPair(pubKey, prvKey);
     }
 
     /**
      * <P>
-     * ASN.1 syntax according to rfc5915 is:
+     * ASN.1 syntax according to <A HREF="https://tools.ietf.org/html/rfc5915">RFC 5915</A> is:
      * </P>
      * </BR>
-     * 
+     *
      * <PRE>
      * <CODE>
      * ECPrivateKey ::= SEQUENCE {
@@ -109,7 +134,7 @@ public class ECDSAPEMResourceKeyPairParser extends AbstractPEMResourceKeyPairPar
      * <I>ECParameters</I> syntax according to RFC5480:
      * </P>
      * </BR>
-     * 
+     *
      * <PRE>
      * <CODE>
      * ECParameters ::= CHOICE {
@@ -119,47 +144,55 @@ public class ECDSAPEMResourceKeyPairParser extends AbstractPEMResourceKeyPairPar
      * }
      * </CODE>
      * </PRE>
-     * 
-     * @param  inputStream The {@link InputStream} containing the DER encoded data
-     * @param  okToClose   {@code true} if OK to close the DER stream once parsing complete
+     *
+     * @param  curve       The {@link ECCurves curve} represented by this data (in case it was optional and somehow
+     *                     known externally) if {@code null} then it is assumed to be part of the parsed data.
+     * @param  sequence    The {@link ASN1Object} sequence containing the DER encoded data
      * @return             The decoded {@link SimpleImmutableEntry} of {@link ECPublicKeySpec} and
      *                     {@link ECPrivateKeySpec}
      * @throws IOException If failed to to decode the DER stream
      */
-    public static SimpleImmutableEntry<ECPublicKeySpec, ECPrivateKeySpec> decodeECPrivateKeySpec(
-            InputStream inputStream, boolean okToClose)
+    public static Map.Entry<ECPublicKeySpec, ECPrivateKeySpec> decodeECPrivateKeySpec(ECCurves curve, ASN1Object sequence)
             throws IOException {
-        ASN1Object sequence;
-        try (DERParser parser = new DERParser(NoCloseInputStream.resolveInputStream(inputStream, okToClose))) {
-            sequence = parser.readObject();
+        ASN1Type objType = (sequence == null) ? null : sequence.getObjType();
+        if (!ASN1Type.SEQUENCE.equals(objType)) {
+            throw new IOException("Invalid DER: not a sequence: " + objType);
         }
 
-        if (!ASN1Type.SEQUENCE.equals(sequence.getObjType())) {
-            throw new IOException("Invalid DER: not a sequence: " + sequence.getObjType());
-        }
-
-        // Parse inside the sequence
         try (DERParser parser = sequence.createParser()) {
-            ECPrivateKeySpec prvSpec = decodeECPrivateKeySpec(parser);
-            ECCurves curve = ECCurves.fromCurveParameters(prvSpec.getParams());
-            if (curve == null) {
-                throw new StreamCorruptedException("Unknown curve");
-            }
-
-            ECPoint w = decodeECPublicKeyValue(curve, parser);
+            Map.Entry<ECPrivateKeySpec, ASN1Object> result = decodeECPrivateKeySpec(curve, parser);
+            ECPrivateKeySpec prvSpec = result.getKey();
+            ASN1Object publicData = result.getValue();
+            ECPoint w = (publicData == null) ? decodeECPublicKeyValue(parser) : decodeECPointData(publicData);
             ECPublicKeySpec pubSpec = new ECPublicKeySpec(w, prvSpec.getParams());
             return new SimpleImmutableEntry<>(pubSpec, prvSpec);
         }
     }
 
-    public static final ECPrivateKeySpec decodeECPrivateKeySpec(DERParser parser) throws IOException {
+    /*
+     * According to https://tools.ietf.org/html/rfc5915 - section 3
+     *
+     * ECPrivateKey ::= SEQUENCE {
+     *      version INTEGER { ecPrivkeyVer1(1) } (ecPrivkeyVer1),
+     *      privateKey OCTET STRING,
+     *      parameters [0] ECParameters {{ NamedCurve }} OPTIONAL,
+     *      publicKey [1] BIT STRING OPTIONAL
+     * }
+     */
+    public static Map.Entry<ECPrivateKeySpec, ASN1Object> decodeECPrivateKeySpec(ECCurves curve, DERParser parser)
+            throws IOException {
         // see openssl asn1parse -inform PEM -in ...file... -dump
-        ASN1Object versionObject = parser.readObject(); // Skip version
+        ASN1Object versionObject = parser.readObject();
         if (versionObject == null) {
             throw new StreamCorruptedException("No version");
         }
 
-        // as per RFC-5915 section 3
+        /*
+         * According to https://tools.ietf.org/html/rfc5915 - section 3
+         *
+         * For this version of the document, it SHALL be set to ecPrivkeyVer1,
+         * which is of type INTEGER and whose value is one (1)
+         */
         BigInteger version = versionObject.asInteger();
         if (!BigInteger.ONE.equals(version)) {
             throw new StreamCorruptedException("Bad version value: " + version);
@@ -175,18 +208,60 @@ public class ECDSAPEMResourceKeyPairParser extends AbstractPEMResourceKeyPairPar
             throw new StreamCorruptedException("Non-matching private key object type: " + objType);
         }
 
-        ASN1Object paramsObject = parser.readObject();
-        if (paramsObject == null) {
-            throw new StreamCorruptedException("No parameters value");
+        /*
+         * According to https://tools.ietf.org/html/rfc5915 - section 3
+         *
+         * parameters specifies the elliptic curve domain parameters associated to the private key. The type
+         * ECParameters is discussed in [RFC5480]. As specified in [RFC5480], only the namedCurve CHOICE is permitted.
+         * namedCurve is an object identifier that fully identifies the required values for a particular set of elliptic
+         * curve domain parameters. Though the ASN.1 indicates that the parameters field is OPTIONAL, implementations
+         * that conform to this document MUST always include the parameters field.
+         */
+        Map.Entry<ECCurves, ASN1Object> result = parseCurveParameter(parser);
+        ECCurves namedParam = (result == null) ? null : result.getKey();
+        if (namedParam == null) {
+            if (curve == null) {
+                throw new StreamCorruptedException("Cannot determine curve type");
+            }
+        } else if (curve == null) {
+            curve = namedParam;
+        } else if (namedParam != curve) {
+            throw new StreamCorruptedException("Mismatched provide (" + curve + ") vs. parsed curve (" + namedParam + ")");
         }
 
-        // TODO make sure params object tag is 0xA0
+        BigInteger s = ECCurves.octetStringToInteger(keyObject.getPureValueBytes());
+        ECPrivateKeySpec keySpec = new ECPrivateKeySpec(s, curve.getParameters());
+        return new SimpleImmutableEntry<>(keySpec, (result == null) ? null : result.getValue());
+    }
+
+    public static Map.Entry<ECCurves, ASN1Object> parseCurveParameter(DERParser parser) throws IOException {
+        return parseCurveParameter(parser.readObject());
+    }
+
+    public static Map.Entry<ECCurves, ASN1Object> parseCurveParameter(ASN1Object paramsObject) throws IOException {
+        if (paramsObject == null) {
+            return null;
+        }
+
+        ASN1Type objType = paramsObject.getObjType();
+        if (objType == ASN1Type.NULL) {
+            return null;
+        }
 
         List<Integer> curveOID;
         try (DERParser paramsParser = paramsObject.createParser()) {
             ASN1Object namedCurve = paramsParser.readObject();
             if (namedCurve == null) {
                 throw new StreamCorruptedException("Missing named curve parameter");
+            }
+
+            /*
+             * The curve OID is OPTIONAL - if it is not there then the
+             * public key data replaces it
+             */
+            objType = namedCurve.getObjType();
+            if (objType == ASN1Type.BIT_STRING) {
+                return new SimpleImmutableEntry<>(null, namedCurve);
             }
 
             curveOID = namedCurve.asOID();
@@ -197,8 +272,7 @@ public class ECDSAPEMResourceKeyPairParser extends AbstractPEMResourceKeyPairPar
             throw new StreamCorruptedException("Unknown curve OID: " + curveOID);
         }
 
-        BigInteger s = ECCurves.octetStringToInteger(keyObject.getPureValueBytes());
-        return new ECPrivateKeySpec(s, curve.getParameters());
+        return new SimpleImmutableEntry<>(curve, null);
     }
 
     /**
@@ -206,40 +280,50 @@ public class ECDSAPEMResourceKeyPairParser extends AbstractPEMResourceKeyPairPar
      * ASN.1 syntax according to rfc5915 is:
      * </P>
      * </BR>
-     * 
+     *
      * <pre>
      * <code>
      *      publicKey  [1] BIT STRING OPTIONAL
      * </code>
      * </pre>
-     * 
-     * @param  curve       The {@link ECCurves} curve
+     *
      * @param  parser      The {@link DERParser} assumed to be positioned at the start of the data
      * @return             The encoded {@link ECPoint}
      * @throws IOException If failed to create the point
      */
-    public static final ECPoint decodeECPublicKeyValue(ECCurves curve, DERParser parser) throws IOException {
+    public static final ECPoint decodeECPublicKeyValue(DERParser parser) throws IOException {
+        return decodeECPublicKeyValue(parser.readObject());
+    }
+
+    public static final ECPoint decodeECPublicKeyValue(ASN1Object dataObject) throws IOException {
         // see openssl asn1parse -inform PEM -in ...file... -dump
-        ASN1Object dataObject = parser.readObject();
         if (dataObject == null) {
             throw new StreamCorruptedException("No public key data bytes");
         }
 
+        /*
+         * According to https://tools.ietf.org/html/rfc5915
+         *
+         * Though the ASN.1 indicates publicKey is OPTIONAL, implementations
+         * that conform to this document SHOULD always include the publicKey field
+         */
         try (DERParser dataParser = dataObject.createParser()) {
-            ASN1Object pointData = dataParser.readObject();
-            if (pointData == null) {
-                throw new StreamCorruptedException("Missing public key data parameter");
-            }
-
-            ASN1Type objType = pointData.getObjType();
-            if (!ASN1Type.BIT_STRING.equals(objType)) {
-                throw new StreamCorruptedException("Non-matching public key object type: " + objType);
-            }
-
-            // see https://tools.ietf.org/html/rfc5480#section-2.2
-            byte[] octets = pointData.getValue();
-            return ECCurves.octetStringToEcPoint(octets);
+            return decodeECPointData(dataParser.readObject());
         }
     }
 
+    public static final ECPoint decodeECPointData(ASN1Object pointData) throws IOException {
+        if (pointData == null) {
+            throw new StreamCorruptedException("Missing public key data parameter");
+        }
+
+        ASN1Type objType = pointData.getObjType();
+        if (!ASN1Type.BIT_STRING.equals(objType)) {
+            throw new StreamCorruptedException("Non-matching public key object type: " + objType);
+        }
+
+        // see https://tools.ietf.org/html/rfc5480#section-2.2
+        byte[] octets = pointData.getValue();
+        return ECCurves.octetStringToEcPoint(octets);
+    }
 }