+++ sshd-common/src/test/java/org/apache/sshd/common/config/keys/loader/pem/PKCS8PEMResourceKeyPairParserTest.java
@@ -21,6 +21,7 @@ package org.apache.sshd.common.config.keys.loader.pem;
 import java.io.ByteArrayInputStream;
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.net.URL;
 import java.security.GeneralSecurityException;
 import java.security.KeyPair;
 import java.security.KeyPairGenerator;
@@ -32,12 +33,14 @@ import java.util.List;
 import org.apache.commons.ssl.PEMItem;
 import org.apache.commons.ssl.PEMUtil;
 import org.apache.sshd.common.NamedResource;
+import org.apache.sshd.common.cipher.ECCurves;
 import org.apache.sshd.common.config.keys.KeyUtils;
 import org.apache.sshd.common.util.GenericUtils;
 import org.apache.sshd.common.util.security.SecurityUtils;
 import org.apache.sshd.util.test.JUnit4ClassRunnerWithParametersFactory;
 import org.apache.sshd.util.test.JUnitTestSupport;
 import org.apache.sshd.util.test.NoIoTestCase;
+import org.junit.Assume;
 import org.junit.FixMethodOrder;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -65,7 +68,7 @@ public class PKCS8PEMResourceKeyPairParserTest extends JUnitTestSupport {
         this.keySize = keySize;
     }
 
-    @Parameters(name = "{0} / {1}")
+    @Parameters(name = "{0}-{1}")
     public static List<Object[]> parameters() {
         List<Object[]> params = new ArrayList<>();
         for (Integer ks : RSA_SIZES) {
@@ -74,17 +77,30 @@ public class PKCS8PEMResourceKeyPairParserTest extends JUnitTestSupport {
         for (Integer ks : DSS_SIZES) {
             params.add(new Object[] { KeyUtils.DSS_ALGORITHM, ks });
         }
+        if (SecurityUtils.isECCSupported()) {
+            for (ECCurves curve : ECCurves.VALUES) {
+                if (!curve.isSupported()) {
+                    outputDebugMessage("Skip unsupported curve=%s", curve);
+                    continue;
+                }
+
+                params.add(new Object[] { KeyUtils.EC_ALGORITHM, curve.getKeySize() });
+            }
+        }
+        if (SecurityUtils.isEDDSACurveSupported()) {
+            params.add(new Object[] { SecurityUtils.EDDSA, 0 });
+        }
         return params;
     }
 
     @Test // see SSHD-760
-    public void testPkcs8() throws IOException, GeneralSecurityException {
+    public void testLocallyGeneratedPkcs8() throws IOException, GeneralSecurityException {
         KeyPairGenerator generator = SecurityUtils.getKeyPairGenerator(algorithm);
         if (keySize > 0) {
             generator.initialize(keySize);
         }
-        KeyPair kp = generator.generateKeyPair();
 
+        KeyPair kp = generator.generateKeyPair();
         try (ByteArrayOutputStream os = new ByteArrayOutputStream()) {
             Collection<Object> items = new ArrayList<>();
             PrivateKey prv1 = kp.getPrivate();
@@ -102,6 +118,30 @@ public class PKCS8PEMResourceKeyPairParserTest extends JUnitTestSupport {
                 assertKeyEquals("Mismatched private key", prv1, kp2.getPrivate());
             }
         }
+    }
+
+    /*
+     * See https://gist.github.com/briansmith/2ee42439923d8e65a266994d0f70180b
+     *
+     * openssl genpkey -algorithm RSA -pkeyopt rsa_keygen_bits:1024 -out pkcs8-rsa-1024.pem
+     * openssl genpkey -algorithm EC -pkeyopt ec_paramgen_curve:P-256 -pkeyopt ec_param_enc:named_curve -out pkcs8-ecdsa-256.pem
+     *
+     * openssl ecparam -genkey -name prime256v1 -noout -out pkcs8-ec-256.key
+     * openssl pkcs8 -topk8 -inform PEM -outform PEM -nocrypt -in pkcs8-ec-256.key -out pkcs8-ec-256.pem
+     *
+     * openssl genpkey -algorithm ed25519 -out pkcs8-ed25519.pem
+     * openssl asn1parse -inform PEM -in ...file... -dump
+     */
+    @Test // see SSHD-989
+    public void testPKCS8FileParsing() throws Exception {
+        String baseName = "pkcs8-" + algorithm.toLowerCase();
+        String resourceKey = baseName + ((keySize > 0) ? "-" + keySize : "") + ".pem";
+        URL url = getClass().getResource(resourceKey);
+        Assume.assumeTrue("No test file=" + resourceKey, url != null);
+
+        Collection<KeyPair> pairs = PKCS8PEMResourceKeyPairParser.INSTANCE.loadKeyPairs(null, url, null);
+        assertEquals("Mismatched extract keys count", 1, GenericUtils.size(pairs));
+        validateKeyPairSignable(algorithm + "/" + keySize, GenericUtils.head(pairs));
     }
 
     @Override